creacion nueva tabla y población (menu)
