--- src/main/java/co/edu/escuelaing/cvds/lab7/controller/PedidoController.java
@@ -1,16 +1,0 @@
-package co.edu.escuelaing.cvds.lab7.controller;
-import co.edu.escuelaing.cvds.lab7.model.Menu;
-import co.edu.escuelaing.cvds.lab7.service.PlatoService;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Controller;
-import org.springframework.ui.Model;
-import org.springframework.web.bind.annotation.GetMapping;
-import java.util.List;
-
-@Controller
-public class PedidoController {
-        @GetMapping("/pedido")
-        public String mostrarPaginaPedido() {
-            return "pedido";
-        }
-    }
