+++ src/main/java/co/edu/escuelaing/cvds/lab7/repository/MenuRepository.java
@@ -1,13 +1,10 @@
 package co.edu.escuelaing.cvds.lab7.repository;
-
 import co.edu.escuelaing.cvds.lab7.model.Menu;
 import co.edu.escuelaing.cvds.lab7.model.User;
 import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
 
 import java.util.List;
 
-@Repository
-public interface PaisRepository extends JpaRepository <User, String> {
-    public List<User> findByUsername(String username);
+public interface MenuRepository extends JpaRepository<Menu, Integer> {
+    public List<Menu> findByCaducidad(String caducidad);
 }
