--- src/main/java/co/edu/escuelaing/cvds/lab7/repository/PlatoRepository.java
@@ -1,7 +1,0 @@
-package co.edu.escuelaing.cvds.lab7.repository;
-import co.edu.escuelaing.cvds.lab7.model.Menu;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface PlatoRepository extends JpaRepository<Menu, String> {
-    // Puedes agregar consultas personalizadas aquí si es necesario
-}
