--- src/main/java/co/edu/escuelaing/cvds/lab7/service/PlatoService.java
@@ -1,17 +1,0 @@
-package co.edu.escuelaing.cvds.lab7.service;
-import co.edu.escuelaing.cvds.lab7.model.Menu;
-import co.edu.escuelaing.cvds.lab7.repository.PlatoRepository;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Service;
-
-import java.util.List;
-
-@Service
-public class PlatoService {
-    @Autowired
-    private PlatoRepository platoRepository;
-
-    public List<Menu> obtenerTodosLosPlatos() {
-        return platoRepository.findAll();
-    }
-}
